langchain_rag/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.0.10
langchain-postgres==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
faiss-cpu==1.7.4
google-generativeai==0.3.2
numpy==1.24.3
PyPDF2==3.0.1
